fixed amount overflow u64 to u128 ic20 packet in ibc.rs of cw20-ics20

+++ contracts/cw20-ics20/src/ibc.rs
@@ -56,7 +56,7 @@ impl Ics20Packet {
     }
 
     pub fn validate(&self) -> Result<(), ContractError> {
-        if self.amount.u128() > (u64::MAX as u128) {
+        if self.amount.u128() > (u128::MAX as u128) {
             Err(ContractError::AmountOverflow {})
         } else {
             Ok(())
